python/requirements-kafka-python.txt
kafka-python==2.0.2
asyncio==3.4.3